+++ packages/client/src/StreamVideoClient.ts
@@ -11,6 +11,8 @@ import type {
   CreateGuestResponse,
   GetEdgesResponse,
   ListDevicesResponse,
+  QueryAggregateCallStatsRequest,
+  QueryAggregateCallStatsResponse,
   QueryCallsRequest,
   QueryCallsResponse,
   QueryCallStatsRequest,
@@ -263,7 +265,6 @@ export class StreamVideoClient {
           );
           return;
         }
-
         this.logger('info', `New call created and registered: ${call.cid}`);
         const newCall = new Call({
           streamClient: this.streamClient,
@@ -287,25 +288,24 @@ export class StreamVideoClient {
           );
           return;
         }
-
-        // The call might already be tracked by the client,
         // if `call.created` was received before `call.ring`.
-        // In that case, we cleanup the already tracked call.
-        const prevCall = this.writeableStateStore.findCall(call.type, call.id);
-        await prevCall?.leave({ reason: 'cleaning-up in call.ring' });
-        // we create a new call
-        const theCall = new Call({
-          streamClient: this.streamClient,
-          type: call.type,
-          id: call.id,
-          members,
-          clientStore: this.writeableStateStore,
-          ringing: true,
-        });
-        theCall.state.updateFromCallResponse(call);
-        // we fetch the latest metadata for the call from the server
-        await theCall.get();
-        this.writeableStateStore.registerCall(theCall);
+        // the client already has the call instance and we just need to update the state
+        const theCall = this.writeableStateStore.findCall(call.type, call.id);
+        if (theCall) {
+          await theCall.updateFromRingingEvent(event);
+        } else {
+          // if client doesn't have the call instance, create the instance and fetch the latest state
+          // Note: related - we also have onRingingCall method to handle this case from push notifications
+          const newCallInstance = new Call({
+            streamClient: this.streamClient,
+            type: call.type,
+            id: call.id,
+            members,
+            clientStore: this.writeableStateStore,
+            ringing: true,
+          });
+          await newCallInstance.get();
+        }
       }),
     );
 
@@ -439,6 +439,21 @@ export class StreamVideoClient {
       QueryCallStatsResponse,
       QueryCallStatsRequest
     >(`/call/stats`, data);
+  };
+
+  /**
+   * Retrieve the list of available reports aggregated from the call stats.
+   *
+   * @param data Specify filter conditions like from and to (within last 30 days) and the report types
+   * @returns Requested reports with (mostly) raw daily data for each report type requested
+   */
+  queryAggregateCallStats = async (
+    data: QueryAggregateCallStatsRequest = {},
+  ) => {
+    return this.streamClient.post<
+      QueryAggregateCallStatsResponse,
+      QueryAggregateCallStatsRequest
+    >(`/stats`, data);
   };
 
   /**